+++ example/lib/sub/child.dart
@@ -1,0 +1,7 @@
+import 'package:more_visibility_example/visibility_root.dart';
+import 'internal.dart';
+
+final okProtected = protectedValue;
+final badDefault = defaultValue; // expect lint: @mdefault from parent dir
+
+final okSubProtected = subProtected;