+++ example/lib/sub/internal.dart
@@ -1,0 +1,7 @@
+import 'package:more_visibility/annotations.dart';
+
+@mprotected
+const subProtected = 'sub';
+
+@mdefault
+const subDefault = 'sub-default';