+++ example/lib/visibility_root.dart
@@ -1,0 +1,9 @@
+import 'package:more_visibility/annotations.dart';
+
+final visibleEverywhere = 0;
+
+@mprotected
+final protectedValue = 1;
+
+@mdefault
+final defaultValue = 2;