+++ lib/more_visibility.dart
@@ -1,0 +1,10 @@
+library more_visibility;
+
+export 'annotations.dart';
+
+import 'package:custom_lint_builder/custom_lint_builder.dart';
+
+import 'src/plugin.dart';
+
+/// Entrypoint for the custom_lint plugin.
+PluginBase createPlugin() => MoreVisibilityPlugin();